Tidy Up (#16)
* Remove unused Git hooks and migrated config

Deleted numerous unused Git hook sample files and log references to clean up the repository. Also added an empty "dependencies" object to the package.json file to reflect the current project configuration.

* Add new yarn.lock file to project

This commit includes the autogenerated file `yarn.lock` that contains the dependency tree and locking specific versions for consistency. The file should not be edited manually as it ensures the exact versions of dependencies are installed.